+++ lib/src/printer_manager.dart
@@ -2,10 +2,12 @@ import 'dart:io';
 
 import 'package:flutter_pos_printer_platform_image_3/flutter_pos_printer_platform_image_3.dart';
 
-enum PrinterType { bluetooth, usb, network }
+/// USB-only fork (2026-09-11): `bluetooth` has been dropped from this enum.
+/// Bluetooth support was removed from the plugin entirely -- see the header of
+/// FlutterPosPrinterPlatformPlugin.kt.
+enum PrinterType { usb, network }
 
 class PrinterManager {
-  final bluetoothPrinterConnector = BluetoothPrinterConnector.instance;
   final tcpPrinterConnector = TcpPrinterConnector.instance;
   final usbPrinterConnector = UsbPrinterConnector.instance;
 
@@ -15,10 +17,8 @@ class PrinterManager {
 
   static PrinterManager get instance => _instance;
 
-  Stream<PrinterDevice> discovery({required PrinterType type, bool isBle = false, TcpPrinterInput? model}) {
-    if (type == PrinterType.bluetooth && (Platform.isIOS || Platform.isAndroid)) {
-      return bluetoothPrinterConnector.discovery(isBle: isBle);
-    } else if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
+  Stream<PrinterDevice> discovery({required PrinterType type, TcpPrinterInput? model}) {
+    if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
       return usbPrinterConnector.discovery();
     } else {
       return tcpPrinterConnector.discovery(model: model);
@@ -26,14 +26,7 @@ class PrinterManager {
   }
 
   Future<bool> connect({required PrinterType type, required BasePrinterInput model}) async {
-    if (type == PrinterType.bluetooth && (Platform.isIOS || Platform.isAndroid)) {
-      try {
-        var conn = await bluetoothPrinterConnector.connect(model as BluetoothPrinterInput);
-        return conn;
-      } catch (e) {
-        throw Exception('model must be type of BluetoothPrinterInput');
-      }
-    } else if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
+    if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
       try {
         var conn = await usbPrinterConnector.connect(model as UsbPrinterInput);
         return conn;
@@ -51,9 +44,7 @@ class PrinterManager {
   }
 
   Future<bool> disconnect({required PrinterType type, int? delayMs}) async {
-    if (type == PrinterType.bluetooth && (Platform.isIOS || Platform.isAndroid)) {
-      return await bluetoothPrinterConnector.disconnect();
-    } else if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
+    if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
       return await usbPrinterConnector.disconnect(delayMs: delayMs);
     } else {
       return await tcpPrinterConnector.disconnect();
@@ -61,18 +52,14 @@ class PrinterManager {
   }
 
   Future<bool> send({required PrinterType type, required List<int> bytes}) async {
-    if (type == PrinterType.bluetooth && (Platform.isIOS || Platform.isAndroid)) {
-      return await bluetoothPrinterConnector.send(bytes);
-    } else if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
+    if (type == PrinterType.usb && (Platform.isAndroid || Platform.isWindows)) {
       return await usbPrinterConnector.send(bytes);
     } else {
       return await tcpPrinterConnector.send(bytes);
     }
   }
 
-  Stream<BTStatus> get stateBluetooth => bluetoothPrinterConnector.currentStatus.cast<BTStatus>();
   Stream<USBStatus> get stateUSB => usbPrinterConnector.currentStatus.cast<USBStatus>();
 
-  BTStatus get currentStatusBT => bluetoothPrinterConnector.status;
   USBStatus get currentStatusUSB => usbPrinterConnector.status;
 }